+++ History.md
@@ -1,3 +1,4 @@
 ## 0.1.0 - 2013-08-24 - Intial release
 
   * methods where && query
+  * woth with dynamic collections